Enable suppression of new segments when fractional CWND in use
Added mechanism to tcp_nice_cong_avoid() to support fractional CWND. Support for the use of fractional CWND is emulated by permitting transmission of new segments for a given number of RTTs. The intervals are determined by the denominator (fractional_cwnd), but as TCP transmits new segments in batches of two or more, the interval between the release of new segments will be 2 * fractional_cwnd.