Fixed initialisation of maxRTT
Moved initialisation of maxRTT to tcp_nice_pkts_acked() and ensure that the value is only initialised at the first RTT estimate. Initial value is now based on baseRTT, which should be a closer approximation of the propagation delay and more closely reflect the intent of Venkataramani, Kokku and Dahlin (2002).