Simplify calculation of congestion threshold
Simplified the formula used to calculate the congestion indication threshold. This calculation is already based on RTT in microseconds, which makes the previously added precision redundant.

This change also prevents the calculation from exceeding the maximum possible value of u32, resulting in erroneous threshold values as the calculated value wraps.